sealing: More cleanup
